+++ src/components/Authenticated/ImageMetadata.vue
@@ -12,14 +12,16 @@
 <script lang='ts'>
 
 import Vue from 'vue';
-import { secondsToText } from '@/vanillaTS/secondsToText';
-import { ImageModule, PiStatusModule } from '@/store';
+
 import { convert_bytes } from '@/vanillaTS/convertBytes';
+import { imageModule, piStatusModule } from '@/store';
+import { mapStores } from 'pinia';
 import { mdiClock, mdiUpdate, mdiImageSizeSelectLarge, mdiImage } from '@mdi/js';
+import { nextUpdateToText } from '@/vanillaTS/secondsToText';
+import { TDataToDisplay } from '@/types';
 import DisplayRows from '@/components/Authenticated/DisplayRows.vue';
 import PiOffline from '@/components/Authenticated/PiOffline.vue';
-
-import { TDataToDisplay } from '@/types';
+import { zeroPad } from '@/vanillaTS/zeropad';
 
 export default Vue.extend({
 	name: 'image-metadata-component',
@@ -30,24 +32,26 @@ export default Vue.extend({
 	},
 
 	computed: {
+		...mapStores(imageModule, piStatusModule),
+
 		imageSize_compressed (): number|undefined {
-			return ImageModule.imageSize_compressed;
+			return this.imageStore.imageSize_compressed;
 		},
 		imageSize_original (): number|undefined {
-			return ImageModule.imageSize_original;
+			return this.imageStore.imageSize_original;
 		},
 		init ():boolean {
-			return PiStatusModule.init;
+			return this.piStatusStore.init;
 		},
 		intervalToHMS (): string {
-			return secondsToText(this.updateCountdown*1000);
+			return nextUpdateToText(this.updateCountdown*1000);
 		},
-		ddd (): string {
+		formattedTimestamp (): string {
 			if (!this.timestamp) return '';
 			return this.formatDate(this.timestamp);
 		},
 		piOnline (): boolean {
-			return PiStatusModule.online;
+			return this.piStatusStore.online;
 		},
 		piInfo (): TDataToDisplay {
 			return [
@@ -55,9 +59,7 @@ export default Vue.extend({
 					{
 						icon: mdiClock,
 						text: 'taken',
-						// This is why?
-						// TODO fix this
-						value: this.ddd,
+						value: this.formattedTimestamp,
 					},
 					{
 						icon: mdiUpdate,
@@ -81,10 +83,10 @@ export default Vue.extend({
 			];
 		},
 		timestamp (): Date|undefined {
-			return ImageModule.timestamp ? new Date(ImageModule.timestamp) : undefined;
+			return this.imageStore.timestamp ? new Date(this.imageStore.timestamp) : undefined;
 		},
 		updateCountdown (): number {
-			return ImageModule.updateCountdown;
+			return this.imageStore.updateCountdown;
 		},
 	},
 
@@ -107,12 +109,8 @@ export default Vue.extend({
 		},
 
 		formatDate (data: Date): string {
-			return `${this.dayOptions[data.getDay()]} ${data.getFullYear()}-${this.zeroPad(data.getMonth() + 1)}-${this.zeroPad(data.getDate())} @ ${this.zeroPad(data.getHours())}:${this.zeroPad(data.getMinutes())}:${this.zeroPad(data.getSeconds())}`;
+			return `${this.dayOptions[data.getDay()]} ${data.getFullYear()}-${zeroPad(data.getMonth() + 1)}-${zeroPad(data.getDate())} @ ${zeroPad(data.getHours())}:${zeroPad(data.getMinutes())}:${zeroPad(data.getSeconds())}`;
 		},
-
-		zeroPad (unit: number): string {
-			return String(unit).padStart(2, '0');
-		}
 		
 	},