+++ src/store/index.ts
@@ -1,16 +1,13 @@
-import loading from '@/store/modules/loading';
-import piStatus from '@/store/modules/piStatus';
-import snack from '@/store/modules/snackbar';
-import user from '@/store/modules/user';
-import ws from '@/store/modules/websocket';
-import image from '@/store/modules/image';
+import { imageStore } from '@/store/image';
+import { loadingStore } from '@/store/loading';
+import { piStatusStore } from '@/store/piStatus';
+import { snackbarStore } from '@/store/snackbar';
+import { userStore } from '@/store/user';
+import { websocketStore } from '@/store/websocket';
 
-import vuex from '@/store/vuex_setup';
-export const store = vuex;
-
-export const LoadingModule = loading;
-export const PiStatusModule = piStatus;
-export const SnackModule = snack;
-export const UserModule = user;
-export const WSModule = ws;
-export const ImageModule = image;
+export const imageModule = imageStore;
+export const loadingModule = loadingStore;
+export const piStatusModule = piStatusStore;
+export const snackbarModule = snackbarStore;
+export const userModule = userStore;
+export const websocketModule = websocketStore;