+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+# <a href='https://github.com/mrjackwills/leafcast_vue/releases/tag/v0.7.0'>v0.7.0</a>
+### 2021-12-04
+
+### Chores
++ npm update, [6fda709d6e999be991ef7297fbdbc1db92f59516](https://github.com/mrjackwills/leafcast_vue/commit/6fda709d6e999be991ef7297fbdbc1db92f59516), [9c2bd4f654fbacbabefa099a16153f9d76cb6353](https://github.com/mrjackwills/leafcast_vue/commit/9c2bd4f654fbacbabefa099a16153f9d76cb6353), [08a5c7cc1171afaa919d5db1df2cadbe931a0e3e](https://github.com/mrjackwills/leafcast_vue/commit/08a5c7cc1171afaa919d5db1df2cadbe931a0e3e)
+
+### Features
++ video page/location, for showing timelapses, [864c7b4c900f1ae737f86dab012976cfa466ab4c](https://github.com/mrjackwills/leafcast_vue/commit/864c7b4c900f1ae737f86dab012976cfa466ab4c) (origin/dev), [c1aa84da784e187684badd97534856fe6129ad09](https://github.com/mrjackwills/leafcast_vue/commit/c1aa84da784e187684badd97534856fe6129ad09)
++ replace vuex with pinia, [d1168edbd4a0cc0b7b8bf1970854974fa4d04ef1](https://github.com/mrjackwills/leafcast_vue/commit/d1168edbd4a0cc0b7b8bf1970854974fa4d04ef1)
++ update secondsToString & refactor out zeropad, [851512c63e2d895ba5ddef438bbaf1fc4ab5b57b](https://github.com/mrjackwills/leafcast_vue/commit/851512c63e2d895ba5ddef438bbaf1fc4ab5b57b)
++ mono-numbers class in variables.scss, [7c0e9361dd6df9191a8f99f0f169328cc18d6e25](https://github.com/mrjackwills/leafcast_vue/commit/7c0e9361dd6df9191a8f99f0f169328cc18d6e25)
+
+### Fixes
++ webpack config, remove fibers, [9a294e5b5f10aea34f92c6d26de8813550a15d5b](https://github.com/mrjackwills/leafcast_vue/commit/9a294e5b5f10aea34f92c6d26de8813550a15d5b)
++ release.sh, [c7cb4a7575062a59ec67ed4d3ecc39ece133ba95](https://github.com/mrjackwills/leafcast_vue/commit/c7cb4a7575062a59ec67ed4d3ecc39ece133ba95), [c791b7a8c32998dbaa75a307587693e5011cb294](https://github.com/mrjackwills/leafcast_vue/commit/c791b7a8c32998dbaa75a307587693e5011cb294)
+
+
+
 # <a href='https://github.com/mrjackwills/leafcast_vue/releases/tag/v0.6.2'>v0.6.2</a>
 #### 2021-11-07